Refactor, some test, split code, add more traits
